+++ src/scenes/Home/components/Donate Section/index.jsx
@@ -34,24 +34,20 @@ class MissionSection extends React.Component {
   render() {
     const { loading, donateBlurb } = this.props.home;
 
-    return (
-      <PageSection className="donate-section">
+    return <PageSection className="donate-section">
         <img src="https://source.unsplash.com/AEaTUnvneik/1600x1024" alt="" />
         <div className="caption hv-center">
-          <div className="text-center w-50">
-            {loading && !donateBlurb ? (
-              <h1 className="text-muted">Loading...</h1>
-            ) : (
-              <React.Fragment>
+          <div className="text-center  col-12 col-md-6">
+            {loading && !donateBlurb ? <h1 className="text-muted">
+                Loading...
+              </h1> : <React.Fragment>
                 <h1>{donateBlurb.blurbTitle}</h1>
                 <p>{donateBlurb.blurbContent}</p>
-                <DonateButton className="btn-light"/>
-              </React.Fragment>
-            )}
+                <DonateButton className="btn-light" />
+              </React.Fragment>}
           </div>
         </div>
-      </PageSection>
-    );
+      </PageSection>;
   }
 }