+++ src/scenes/Home/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import VideoHeader from '../../components/Video Header';
+import HomeMasthead from './components/Home Masthead';
 import ProjectsSection from './components/Projects Section';
 import BlogSection from './components/Blog Section';
 import DonateSection from './components/Donate Section';
@@ -10,7 +10,7 @@ class Home extends React.Component {
   render() {
     return (
       <React.Fragment>
-        <VideoHeader />
+        <HomeMasthead/>
         <ProjectsSection />
         <AboutSection />
         <BlogSection />